requirements.txt
# Core dependencies for CyberSentinel SOC Dashboard
elasticsearch==8.11.1
pandas==2.1.4
numpy==1.24.3
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
pyyaml==6.0.1
colorama==0.4.6
tqdm==4.66.1
faker==20.1.0
ipaddress==1.0.23
geoip2==4.7.0
maxminddb==2.2.0
cryptography==41.0.8
scapy==2.5.0
psutil==5.9.6
schedule==1.2.0
watchdog==3.0.0
jsonschema==4.20.0